+++ src/react/components/dropdown/DropDownToggled.js
@@ -438,7 +438,7 @@ export default class DropDownToggled extends PureComponent {
 			'onKeyDown': compose(trigger.props.onKeyDown, ddt.onTriggerKeyDown),
 			'aria-activedescendant': ddt.state.activeDescendant,
 			'aria-haspopup': true,
-			'aria-controls': dropdown.dropdownId,
+			'aria-controls': dropdown.dropdownId
 		});
 
 		const clonedDropdown = React.cloneElement(dropdown, {
@@ -452,7 +452,7 @@ export default class DropDownToggled extends PureComponent {
 			onCloseAnimationEnd: compose(dropdown.onCloseAnimationEnd, ddt.onCloseAnimationEnd),
 			onOpenAnimationEnd: compose(dropdown.onOpenAnimationEnd, ddt.onOpenAnimationEnd),
 			onKeyDown: compose(dropdown.props.onKeyDown, ddt.onDropDownKeyDown),
-			className: dropdown.props.className,
+			className: dropdown.props.className
 		});
 
 		return (
@@ -472,6 +472,7 @@ export default class DropDownToggled extends PureComponent {
 					isTriggerWidthMatched={ddt.props.matchTriggerWidth}
 					isNotResponsive={forceDesktop}
 					onVisible={shouldAnimate(this) ? null : this.onOpenAnimationEnd}
+					qaHook={qaHook && `${qaHook}--positioning`}
 				>
 						{clonedDropdown}
 				</Positioning>